chore: add Github action to build provider
We want to get rid of the provider-build step in CSE, building the
providers in their own repos makes more sense.

This Github Action creates a release for every commit that is named
`v0.0.0-<commit sha>` and attaches a linux/amd64 and darwin/arm64 binary
to it.

This commit also removes all unneeded actions that came from / with the
fork.